+++ lib/pronto/bundler_audit/results/insecure_source.rb
@@ -5,18 +5,21 @@ require_relative "base_result"
 module Pronto
   class BundlerAudit
     module Results
-      # Pronto::BundlerAudit::Results::InsecureSource builds a Pronto::Message
-      # for Bundler::Audit::Scanner::InsecureSource issues.
+      # Pronto::BundlerAudit::Results::InsecureSource is a stand-in for the
+      # ::Pronto::Message object for ::Bundler::Audit::Scanner::InsecureSource
+      # issues.
       class InsecureSource < BaseResult
-        private
-
-        def report_result
-          build_message(
-            message,
-            line: nil,
-            level: :warning)
+        # @return [Symbol]
+        def level
+          :warning
         end
 
+        # @return [NilClass]
+        def line
+          nil
+        end
+
+        # @return [String]
         def message
           "Insecure Source URI found: #{@scan_result.source}"
         end